requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic==0.28.0
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.5
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0
gunicorn==22.0.0
python-dateutil==2.9.0
